Factor out work of ServiceStatusController and fix broken configuration calls
