+++ api/controller.py
@@ -2,7 +2,6 @@ from nose.tools import set_trace
 import json
 import logging
 import sys
-import time
 import urllib
 import urlparse
 import uuid
@@ -96,7 +95,7 @@ from circulation import (
     CirculationAPI,
     DummyCirculationAPI,
 )
-
+from services import ServiceStatus
 
 class CirculationManager(object):
 
@@ -823,12 +822,7 @@ class WorkController(CirculationManagerController):
 
 class ServiceStatusController(CirculationManagerController):
 
-    def __call__(self):
-        conf = Configuration.authentication_policy()
-        username = conf[Configuration.AUTHENTICATION_TEST_USERNAME]
-        password = conf[Configuration.AUTHENTICATION_TEST_PASSWORD]
-
-        template = """<!DOCTYPE HTML>
+    template = """<!DOCTYPE HTML>
 <html lang="en" class="">
 <head>
 <meta charset="utf8">
@@ -840,55 +834,13 @@ class ServiceStatusController(CirculationManagerController):
 </body>
 </html>
 """
-        timings = dict()
 
-        patrons = []
-        def _add_timing(k, x):
-            try:
-                a = time.time()
-                x()
-                b = time.time()
-                result = b-a
-            except Exception, e:
-                result = e
-            if isinstance(result, float):
-                timing = "SUCCESS: %.2fsec" % result
-            else:
-                timing = "FAILURE: %s" % result
-            timings[k] = timing
-
-        def do_patron():
-            patron = self.conf.auth.authenticated_patron(self.conf.db, username, password)
-            patrons.append(patron)
-            if patron:
-                return patron
-            else:
-                raise ValueError("Could not authenticate test patron!")
-
-        _add_timing('Patron authentication', do_patron)
-
-        patron = patrons[0]
-        def do_overdrive():
-            if not self.conf.overdrive:
-                raise ValueError("Overdrive not configured")
-            return self.conf.overdrive.patron_activity(patron, password)
-        _add_timing('Overdrive patron account', do_overdrive)
-
-        def do_threem():
-            if not self.conf.threem:
-                raise ValueError("3M not configured")
-            return self.conf.threem.patron_activity(patron, password)
-        _add_timing('3M patron account', do_threem)
-
-        def do_axis():
-            if not self.conf.axis:
-                raise ValueError("Axis not configured")
-            return self.conf.axis.patron_activity(patron, password)
-        _add_timing('Axis patron account', do_axis)
-
+    def __call__(self):
+        service_status = ServiceStatus(self._db)
+        timings = service_status.loans_status()
         statuses = []
         for k, v in sorted(timings.items()):
             statuses.append(" <li><b>%s</b>: %s</li>" % (k, v))
 
-        doc = template % dict(statuses="\n".join(statuses))
-        return make_response(doc, 200, {"Content-Type": "text/html"})
+        doc = self.template % dict(statuses="\n".join(statuses))
+        return Response(doc, 200, {"Content-Type": "text/html"})
